% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Downloading Redcap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{User's RedCap API token in their .REnviron file}

\item{redcapUrl}{URL to be queried}

\item{redcapReportId}{specifies the ID for the Redcap Report to be returned}
}
\value{
A Redcap Report in the form of a tibble
}
\description{
This function reads in an API token, url and report id from Redcap, downloading the report and returning the contents as a tibble
}
\details{
Reads in Recap API token, queries redcapUrl to return the Redcap Report, returning the contents as a tibble
}
