% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculating paramter p}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{vector of binary values}
}
\value{
The parameter p that maximizes the log-likelihood
}
\description{
This function takes in a vector of binary values to calculate parameter p
}
\details{
calculates the parameter p that maximizes the log-likelihood log(P(p|data))
using a grid-based search with p in steps of 0.001
}
\examples{
data <- c(1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 1, 0)
logLikBernoulli(data)

}
