% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Find the sample size}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{one sample of data}

\item{x2}{two samples of data (optional)}
}
\value{
The minimum sample size needed for a t-test
}
\description{
This function takes either one or two samples of preliminary data (\code{x1}
 or \code{x2}) and return the minimum sample size needed for a t-test under
 conditions:
 (1) The function should calculate the minimum sample size required
 for a t-test under the null hypothesis
 (2) It should determine the sample size needed to achieve 80% power at a
 significance level of alpha = 0.05
}
\details{
Returns minimum sample size needed for a t-test
}
\examples{
x1 <- rnorm(20)  # Example data for the first sample
x2 <- rnorm(30)  # Example data for the second sample
# Calculate the minimum sample size needed for a t-test
minimumN(x1 = x1, x2 = x2)
minimumN(x1 = x1)
}
