% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate Data using Principal Components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame representing the data to be approximated.}

\item{npc}{An integer specifying the number of principal components to use for approximation.}
}
\value{
A matrix representing the approximation to the original data based on the specified number of principal components.
}
\description{
This function returns an approximation to the data \code{x} based on the specified number of principal components (\code{npc}).
}
\details{
This function performs principal component analysis (PCA) on the input data \code{x} and uses the selected principal components to approximate the original data. The approximation is rescaled and centered to match the original data.
}
\examples{
# Generate sample data
set.seed(123)
x <- matrix(rnorm(100), ncol = 5)

# Approximate data using 3 principal components
pcApprox(x, 3)

}
