% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble}
}
\description{
This function standardizes the variable names in a tibble data to "small camel case".
}
\examples{
library(dplyr)
library(janitor)
data <- tibble::tibble(First_Name = c(1, 2, 3), Last_Name = c(4, 5, 6))
standardizeNames(data)

}
