% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculating & plotting survival curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{numerical vector}

\item{time}{numerical vector}
}
\value{
a plot of the survival curve S(t)
}
\description{
This function takes in numercial vectors status and time to calculate the survival curve S(t), returning a plot of the survival curve
}
\details{
calculates and plots the survival curve S(t) given the numerical
vectors for status and time
}
\examples{
data <- read.csv('https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv')

survCurv(data$status, data$time)

}
