% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscaling Vectors}
\usage{
unscale(x)
}
\arguments{
\item{x}{a scaled vector}
}
\value{
the original vector
}
\description{
This function reverses the transformations of scale() for a vector
}
\details{
Reverses the centering/scaling of a vector applied by scale()
}
\examples{
# Example data
data <- c(1, 2, 3, 4, 5)

# Apply scaling
scaled_data <- scale(data)

# Reverse the scaling
unscale(scaled_data)
}
